+++ FinanceManager/TakeOperationsPage.xaml.cs
@@ -66,7 +66,7 @@ public partial class TakeOperationsPage : ContentPage
             DescriptionEditor.Text = "Отсутствует";
         }
 
-        Single.TryParse(ValueEntry.Text, NumberStyles.Any, CultureInfo.InvariantCulture, out Single value);
+        Double.TryParse(ValueEntry.Text, NumberStyles.Any, CultureInfo.InvariantCulture, out Double value);
         var operation = new AccountStats { AccountID = Convert.ToInt32(accounIDLabel.Text), Value = value * -1, Operation = PikerType.SelectedItem?.ToString() ?? "noType", Description = DescriptionEditor.Text, Type = "outcome", date = chosedDate };
         await database.SaveItemAsync(operation);
         await Navigation.PopAsync();
